test: use pointer-sized integer types in tests and add platform flag to test runner
* tests now use pointer-sized integer types (intptr_t / NSUInteger) to avoid truncation/ABI mismatches for retain counts and mutation pointers
* add --platform option to test/run-test and forward it to mulle-sde link-args so tests can be linked for a specific platform

Replaced long/unsigned long with intptr_t/NSUInteger in test/04_property/accessors2.m and test/11_foreach/foreach.m. Updated test/run-test to introduce a platform variable, parse --platform, and call `exekutor mulle-sde test --platform "${platform}" link-args --output-format ld --startup`.

+++ test/04_property/accessors2.m
@@ -15,7 +15,7 @@
 
 int main()
 {
-    long   oldRetainCount;
+    intptr_t   oldRetainCount;
     SEL _cmd = @selector(foo);
     Base *o = [Base new];
     ptrdiff_t offset = ivar_getOffset(class_getInstanceVariable([Base class], "ivar"));
